requirements.txt
openai==1.34.0
chainlit==0.7.700
langchain
langchain_openai
langchain_community
langchain_core
langchain_huggingface
langchain_text_splitters
python-dotenv==1.0.1
qdrant-client
pypdf
tiktoken